// database.types.ts
export interface Database {
// ...
  public: {
    Tables: {
      products: {
        Row: {
          category: string | null
          created_at: string
          date: string | null
          detail: string | null
          id: string
          name: string | null
          price: number | null
          status: boolean
          user_id: string | null
        }
        Insert: {
          category?: string | null
          created_at?: string
          date?: string | null
          detail?: string | null
          id?: string
          name?: string | null
          price?: number | null
          status?: boolean
          user_id?: string | null
        }
        Update: {
          category?: string | null
          created_at?: string
          date?: string | null
          detail?: string | null
          id?: string
          name?: string | null
          price?: number | null
          status?: boolean
          user_id?: string | null
        }
        Relationships: [
          {
            foreignKeyName: "products_user_id_fkey"
            columns: ["user_id"]
            referencedRelation: "users"
            referencedColumns: ["id"]
          }
        ]
      }
      profile: {
        Row: {
          created_at: string
          goal_money: number | null
          id: string
          name: string | null
          user_id: string | null
        }
        Insert: {
          created_at?: string
          goal_money?: number | null
          id?: string
          name?: string | null
          user_id?: string | null
        }
        Update: {
          created_at?: string
          goal_money?: number | null
          id?: string
          name?: string | null
          user_id?: string | null
        }
        Relationships: [
          {
            foreignKeyName: "profile_user_id_fkey"
            columns: ["user_id"]
            referencedRelation: "users"
            referencedColumns: ["id"]
          }
        ]
      }
    }
    Views: {
      [_ in never]: never
    }
    Functions: {
      [_ in never]: never
    }
    Enums: {
      [_ in never]: never
    }
    CompositeTypes: {
      [_ in never]: never
    }
  }
// ...
}